Common_Core

+++ ph/src/main.c
@@ -3,45 +3,59 @@
 /*                                                        :::      ::::::::   */
 /*   main.c                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: dkhalfao <dkhalfao@student.42.fr>          +#+  +:+       +#+        */
+/*   By: khalfaoui47 <khalfaoui47@student.42.fr>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/19 09:37:46 by dkhalfao          #+#    #+#             */
-/*   Updated: 2025/08/13 14:27:41 by dkhalfao         ###   ########.fr       */
+/*   Updated: 2025/08/14 08:15:13 by khalfaoui47      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../philo.h"
 
-void	check_args(int argc, char **argv)
+int	check_args(int argc, char **argv)
 {
 	int		i;
 	long	num;
 
 	i = 1;
 	if (argc < 5 || argc > 6)
-		error_message("[Argument Count error]\n");
+	{
+		write(1, "Argument Count error\n", 22);
+		return 1;
+	}
 	while (i < argc)
 	{
 		num = ft_atoi(argv[i]);
 		if (i == 1 && (num < 1 || num > PHILO_MAX_COUNT))
-			error_message("[Argument error]\n");
+		{
+			write(1, "Argument error\n", 16);
+			return 1;
+		}
 		else if ((num < 0 || num > INT_MAX))
-			error_message("[Argument error]\n");
+		{
+			write(1, "Argument error\n", 16);
+			return 1;
+		}
 		i++;
 	}
+	return 0;
 }
 
 int	main(int argc, char **argv)
 {
 	t_philo		philos[PHILO_MAX_COUNT];
-	t_mutex		forks[PHILO_MAX_COUNT];
+	pthread_mutex_t		forks[PHILO_MAX_COUNT];
 	t_data	data;
 
-	check_args(argc, argv);
-	init_data(&data, philos, forks);
-	init_forks(&data, forks, ft_atoi(argv[1]));
+	if(check_args(argc, argv))
+		return 1;
+	if(init_data(&data, philos, forks))
+		return 1;
+	if(init_forks(&data, forks, ft_atoi(argv[1])))
+		return 1;
 	init_philos(&data, philos, forks, argv);
-	launcher(&data, philos[0].philo_count);
+	if(simulation(&data, philos[0].philo_count))
+		return 1;
 	destroy_all(&data, NULL, philos[0].philo_count, 0);
 	return (0);
 }
